requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aniso8601==3.0.2          # via flask-restful
click==6.7                # via flask
flask-restful==0.3.6
flask==1.0.2
gunicorn==19.9.0
itsdangerous==0.24        # via flask
jinja2==2.10              # via flask
markupsafe==1.0           # via jinja2
python-dotenv==0.9.1
pytz==2018.5              # via flask-restful
six==1.11.0               # via flask-restful
werkzeug==0.14.1          # via flask
